test(webhook): move runner tests to owning modules
